Refactor VS Code CLI E2E helpers
